rails

+++ app/controllers/shopping_centers_controller.rb
@@ -4,26 +4,42 @@ class ShoppingCentersController < ApplicationController
 
   def show
     @shopping_center=ShoppingCenter.find(params[:id])
-   
-    #ここから閲覧履歴を保存するコード
+    if logged_in?
+    
+      #ここから閲覧履歴を保存するコード
+      new_history = @shopping_center.histories.new
+      new_history.user_id = current_user.id
+  
+      #if文追加、ログイン中のユーザー(current_user)の閲覧履歴(browsing_histories)の中で記事ID(article_id)が
+      #URLに入力されているID(params[:id])と同じものがすでに存在しているか？(exists?)
+      if current_user.histories.exists?(shopping_center_id: "#{params[:id]}")
+        old_history = current_user.histories.find_by(shopping_center_id: "#{params[:id]}")
+        old_history.destroy
+      end
+  
+      new_history.save
+    end
+  end
+  
+  
+def shopping_center_history #閲覧履歴
+  if logged_in?
+    
+      #ここから閲覧履歴を保存するコード
     new_history = @shopping_center.histories.new
     new_history.user_id = current_user.id
-
-    #if文追加、ログイン中のユーザー(current_user)の閲覧履歴(browsing_histories)の中で記事ID(article_id)が
-    #URLに入力されているID(params[:id])と同じものがすでに存在しているか？(exists?)
+  
+      #if文追加、ログイン中のユーザー(current_user)の閲覧履歴(browsing_histories)の中で記事ID(article_id)が
+      #URLに入力されているID(params[:id])と同じものがすでに存在しているか？(exists?)
     if current_user.histories.exists?(shopping_center_id: "#{params[:id]}")
       old_history = current_user.histories.find_by(shopping_center_id: "#{params[:id]}")
       old_history.destroy
     end
-
+  
     new_history.save
-
   end
-  
-  
-  def shopping_center_history #閲覧履歴
-    @history = History.all
-  end
+  @history = current_user.histories.all
+end
   
   
   def search
@@ -36,8 +52,14 @@ class ShoppingCentersController < ApplicationController
     [params[:name1], params[:name2], params[:name3]].reject(&:blank?)
     
   end
-  
-  
+ 
+  def current_user
+    @current_user ||= User.find_by(id: session[:user_id])
+  end
+
+  def logged_in?
+    !!current_user
+  end
   
  
 end
